feat(budget-app): break BudgetTab into many components
