+++ budget-app/src/app/budget/components/types.ts
@@ -1,0 +1,15 @@
+export interface BudgetItem {
+    name: string;
+    assigned: number;
+    activity: number;
+    available: number;
+  }
+  
+  export interface BudgetGroupItem {
+    name: string;
+    budgetItems: BudgetItem[];
+  }
+  
+  export interface BudgetTabProps {
+    items: BudgetGroupItem[];
+  }
